Revert "Merge pull request #1 from midchildan/feat-byte-count"
This reverts commit e8ebb328c611736a4a5c9b641e5ca6c3277cb6f5, reversing
changes made to caac02d4536cb544d20fa379ee348964c4204ab6.

+++ src/nettest_bsd.h
@@ -70,7 +70,7 @@ struct  omni_request_struct {
 
   float      cpu_rate;       /* do we know how fast the cpu is already? */
 
-  int64_t    test_length;    /* how long is the test? */
+  int32_t    test_length;    /* how long is the test? */
 
   uint32_t   so_rcvavoid;    /* avoid copies on recv? */
   uint32_t   so_sndavoid;    /* avoid copies on send? */
